livenewsai/requirements.txt
fastapi>=0.104,<0.110
uvicorn[standard]>=0.24,<0.30
pydantic>=2.9,<3.0
pydantic-settings>=2.4,<3.0
pathway==0.16.1
openai>=1.3,<2.0
requests>=2.31,<3.0
numpy>=1.26,<2.0
python-dotenv>=1.0,<2.0

# Test dependencies (needed for test_livenewsai.py)
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
